html/browsers/browsing-the-web/back-forward-cache/eligibility/dedicated-worker.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="../resources/helper.sub.js"></script>
<script>

promise_test(async t => {
  const pageA = new RemoteContext(token());
  const pageB = new RemoteContext(token());

  const urlA = location.origin + executorPath + pageA.context_id;
  const urlB = originCrossSite + executorPath + pageB.context_id;

  window.open(urlA, '_blank', 'noopener');

  await pageA.execute_script(waitForPageShow);

  assert_equals(await pageA.execute_script(
    () => {
      return new Promise((resolve, reject) => {
        window.worker = new Worker('../resources/echo-worker.js');
        window.worker.onerror = reject;
        window.worker.addEventListener('message', e => resolve(e.data), {once: true});
        window.worker.postMessage('Start');
      });
    }), 'Start');

  await pageA.execute_script(
    (url) => {
      prepareNavigation(() => { location.href = url; });
    },
    [urlB]);

  await pageB.execute_script(waitForPageShow);
  await pageB.execute_script(
    () => {
      prepareNavigation(() => { history.back(); });
    }
  );
  await pageA.execute_script(waitForPageShow);
  await assert_bfcached(pageA);

  // Confirm that the worker is still there.
  assert_equals(await pageA.execute_script(
    () => {
      return new Promise((resolve, reject) => {
        window.worker.onerror = reject;
        window.worker.addEventListener('message', e => resolve(e.data), {once: true});
        window.worker.postMessage('Restored');
      });
    }), 'Restored');

}, 'Eligibility: dedicated workers');
</script>
